refactor: update image system
