+++ src/components/posts/PostCard.astro
@@ -1,6 +1,6 @@
 ---
-import { getImageData } from '$data/posts/images';
 import { Image } from '@astrojs/image/components';
+import { getImageData } from '$helpers/images';
 import { Post, PostType } from '$data/posts';
 import Tag from './Tag.svelte';
