+++ src/images/tauri-svelte.png
