+++ src/images/upgrading-sveltekit.png
