+++ src/layouts/Global.astro
@@ -1,6 +1,6 @@
 ---
 import 'ghostsui';
-import { getImageSrc } from '$data/posts/images';
+import { getImageSrc } from '$helpers/images';
 
 const frontmatter = Astro.props.frontmatter || {};
