+++ src/layouts/Post.astro
@@ -2,8 +2,8 @@
 import SettingsModal from '$components/posts/SettingsModal.svelte';
 import { faTwitter } from '@fortawesome/free-brands-svg-icons';
 import { faRss } from '@fortawesome/free-solid-svg-icons';
-import { getImageSrc } from '$data/posts/images';
 import Tag from '$components/posts/Tag.svelte';
+import { getImageSrc } from '$helpers/images';
 import { tweetLink } from '$helpers/twitter';
 import { format } from 'date-fns';
 import Page from './Page.astro';
